Move meter type API calls to client service
